Make app descriptor use consistent hash deserialization
H256 and H520 already implement `Deserialize` trait. They requires
"0x" prefix that we don't need. We are easy to make a mistake like
ommiting "deserialize_with" attribute.
